add and enable Keccack ARMv7 and ARMv8 NEON support
As both are compiled mutually exclusive, both are reachable via the
sha3*_arm_neon, shake*_arm_neon, cshake*_arm_neon.

ARMv7 was tested on an USB armory MK II. The cSHAKE DRNG 1 GB test with
the C implementation takes about 3 minutes, the Neon support about
53 seconds.

ARMv8 was tested on a Turris Mox I. The cSHAKE DRNG 1 GB test with the
C implementation takes about 24 seconds, the Neon support about 33
seconds. It is strange that the Neon support is so much slower. This
will require further investigation.

+++ aead/tests/cshake_crypt_test.c
@@ -129,7 +129,7 @@ static int cc_tester_cshake_one(const uint8_t *pt, size_t ptlen,
 	if (lc_aead_setkey(cc, key, keylen, NULL, 0))
 		return 1;
 
-	out_enc[0] = (out_enc[0] + 1) &0xff;
+	out_enc[0] = (uint8_t)((out_enc[0] + 1) &0xff);
 	ret = lc_aead_decrypt(cc, out_enc, out_dec, ptlen, aad, aadlen,
 			      tag, exp_tag_len);
 	lc_aead_zero(cc);
