+++ signature/src/dilithium_polyvec.h
@@ -130,7 +130,8 @@ polyvecl_uniform_gamma1(polyvecl *v,
 
 	for (i = 0; i < LC_DILITHIUM_L; ++i)
 		poly_uniform_gamma1(&v->vec[i], seed,
-				    le_bswap16(LC_DILITHIUM_L * nonce + i),
+				    le_bswap16(LC_DILITHIUM_L * nonce +
+					       (uint16_t)i),
 				    ws_buf);
 }
